Use modal & card components and move to modal dir
