+++ src/renderer/modal/about.vue
@@ -1,5 +1,5 @@
 <template>
-	<div v-show="this.$store.state.model.model.about" class="model--medium">
+	<t-card width="medium">
 		<div class="about">
 			<div class="about__header d-flex flex-column align-items-center">
 				<div class="about__header__logo">
@@ -10,22 +10,22 @@
 					{{ appVersion }}
 				</p>
 			</div>
-			<div class="about__content">
-				<p class="about__content__para">
-					Thermal is an electron.js & vue.js based application by which you can
-					use Git with a nice graphic user interface.
-				</p>
-				<p class="about__content__para">
-					This application is copyright and build by CodeCarrot.
-				</p>
-			</div>
 		</div>
-		<div class="about__footer d-flex">
+		<t-card-body class="about__content">
+			<p class="about__content-para">
+				Thermal is an electron.js & vue.js based application by which you can
+				use Git with a nice graphic user interface.
+			</p>
+			<p class="about__content-para">
+				This application is copyright and build by CodeCarrot.
+			</p>
+		</t-card-body>
+		<t-card-footer>
 			<Button
 				text="Close"
 				appearance="outline"
 				margin-left="auto"
-				@click.native="closeModel()"
+				@click.native="closeModal('About')"
 			/>
 			<Button
 				text="Release Notes"
@@ -33,22 +33,34 @@
 				margin-left=".5rem"
 				@click.native="releaseNotes()"
 			/>
-		</div>
-	</div>
+		</t-card-footer>
+	</t-card>
 </template>
 
 <script>
-import thermalLogo from "../icon/logo";
-import packageJson from "../../../../package.json";
-import Button from "../buttons/Button";
+// components
+import TCard from "../components/TCard/TCard";
+import TCardBody from "../components/TCard/TCardBody";
+import TCardFooter from "../components/TCard/TCardFooter";
+import Button from "../components/buttons/Button";
+
+// mixins
+import closeModalMixin from "../mixins/closeModal";
+
+import thermalLogo from "../components/icon/logo";
+import packageJson from "../../../package.json";
 const { shell } = require("electron");
 
 export default {
 	name: "About",
 	components: {
-		thermalLogo,
-		Button
+		TCard,
+		TCardBody,
+		TCardFooter,
+		Button,
+		thermalLogo
 	},
+	mixins: [closeModalMixin],
 	computed: {
 		appVersion() {
 			return `Version ${packageJson.version}`;
@@ -57,9 +69,6 @@ export default {
 	methods: {
 		releaseNotes() {
 			shell.openExternal("https://thermal.codecarrot.net/releases");
-		},
-		closeModel() {
-			this.$store.dispatch("model/showAboutModel");
 		}
 	}
 };
@@ -67,7 +76,6 @@ export default {
 
 <style lang="sass">
 .about
-
 	&__header
 		padding: .8rem 1rem
 		border-bottom: 1px solid #eee
@@ -77,13 +85,8 @@ export default {
 			font-size: .8rem
 
 	&__content
-		padding: .8rem 1rem
+		flex-direction: column
 
-		&__para
+		&-para
 			margin-bottom: .5rem
-
-	&__footer
-		width: 100%
-		border-top: 1px solid #eee
-		padding: .8rem 1rem
 </style>
