+++ src/renderer/modal/addLocalRepository.vue
@@ -1,0 +1,101 @@
+<template>
+	<t-card width="small">
+		<t-card-header>
+			<t-card-header-heading heading="Add local repository" />
+			<t-card-header-close modal-name="AddLocalRepository" />
+		</t-card-header>
+		<t-card-body>
+			<inputText
+				v-model.trim="pathToRepository"
+				v-focus
+				name="pathToRepository"
+				placeholder="Local path to repository"
+				@keyup.esc.native="closeModal('AddLocalRepository')"
+				@keyup.enter.native="addRepository"
+			/>
+			<input
+				ref="fileInput"
+				type="file"
+				name="folderSelector"
+				webkitdirectory
+				style="display: none"
+				@change="fileSectorInput"
+			/>
+			<Button
+				text="Select"
+				appearance="primary"
+				margin-left=".4rem"
+				@click.native="$refs.fileInput.click()"
+			/>
+		</t-card-body>
+		<t-card-footer>
+			<Button
+				text="Close"
+				appearance="outline"
+				margin-left="auto"
+				@click.native="closeModal('AddLocalRepository')"
+			/>
+			<Button
+				appearance="primary"
+				text="Add repository"
+				margin-left=".5rem"
+				@click.native="addRepository(pathToRepository)"
+			/>
+		</t-card-footer>
+	</t-card>
+</template>
+
+<script>
+// components
+import TCard from "../components/TCard/TCard";
+import TCardHeader from "../components/TCard/TCardHeader";
+import TCardHeaderHeading from "../components/TCard/TCardHeaderHeading";
+import TCardHeaderClose from "../components/TCard/TCardHeaderClose";
+import TCardBody from "../components/TCard/TCardBody";
+import TCardFooter from "../components/TCard/TCardFooter";
+import inputText from "../components/input/inputText";
+import Button from "../components/buttons/Button";
+
+// mixins
+import closeModalMixin from "../mixins/closeModal";
+import addRepository from "../mixins/addRepository";
+
+export default {
+	name: "AddLocalRepository",
+	components: {
+		TCard,
+		TCardHeader,
+		TCardHeaderHeading,
+		TCardHeaderClose,
+		TCardBody,
+		TCardFooter,
+		inputText,
+		Button
+	},
+	directives: {
+		focus: {
+			inserted: function(el) {
+				el.focus();
+			}
+		}
+	},
+	mixins: [closeModalMixin, addRepository],
+	data() {
+		return {
+			pathToRepository: ""
+		};
+	},
+	methods: {
+		fileSectorInput(event) {
+			this.pathToRepository = event.target.files[0].path.split("\\").join("/");
+		},
+		addRepository(path) {
+			if (this.pathToRepository !== "") {
+				this.localRepository(path);
+				this.pathToRepository = "";
+				this.closeModal("AddLocalRepository");
+			}
+		}
+	}
+};
+</script>
