+++ src/renderer/modal/initalizeGitRepository.vue
@@ -1,14 +1,12 @@
 <template>
-	<div v-if="isGitRepository" class="model--small isgit">
-		<div class="model__section model__header">
-			<h6 class="model__header__title">
-				Git not found
-			</h6>
-		</div>
-		<div class="model__section model__body">
+	<t-card width="small">
+		<t-card-header>
+			<t-card-header-heading heading="Git not found" />
+		</t-card-header>
+		<t-card-body>
 			By default git will be initalized at the root of the project.
-		</div>
-		<div class="model__section model__footer">
+		</t-card-body>
+		<t-card-footer>
 			<Button
 				text="Go back"
 				appearance="outline"
@@ -22,54 +20,53 @@
 				margin-left=".5rem"
 				@click.native="initalizeGit"
 			/>
-		</div>
-	</div>
+		</t-card-footer>
+	</t-card>
 </template>
 
 <script>
-import Button from "../buttons/Button";
-import gitInit from "../../git/init";
+// components
+import TCard from "../components/TCard/TCard";
+import TCardHeader from "../components/TCard/TCardHeader";
+import TCardHeaderHeading from "../components/TCard/TCardHeaderHeading";
+import TCardBody from "../components/TCard/TCardBody";
+import TCardFooter from "../components/TCard/TCardFooter";
+import Button from "../components/buttons/Button";
+
+// mixins
+import closeModalMixin from "../mixins/closeModal";
+import gitInit from "../git/init";
 
 export default {
 	name: "InitalizeGitRepository",
 	components: {
+		TCard,
+		TCardHeader,
+		TCardHeaderHeading,
+		TCardBody,
+		TCardFooter,
 		Button
 	},
+	mixins: [closeModalMixin],
 	computed: {
 		currentRepository() {
 			return this.$store.getters["workspace/currentRepository"];
-		},
-		isGitRepository() {
-			return (
-				this.$router.history.current.matched[0].path.slice(1) ===
-					"repository" && this.currentRepository.isGit === false
-			);
 		}
 	},
 	methods: {
 		switchRepository() {
-			this.$store.commit("model/toggleModelPlaceholder");
+			this.closeModal("InitalizeGitRepository");
 			this.$store.dispatch("workspace/switchWorkspaceRepository");
 			this.$router.push({ name: "welcome" });
 		},
 		initalizeGit() {
 			gitInit(this.currentRepository.path);
-			this.$store.commit("model/toggleModelPlaceholder");
 			this.$store.dispatch({
 				type: "repository/updateIsGit",
 				isGit: true
 			});
+			this.closeModal("InitalizeGitRepository");
 		}
 	}
 };
 </script>
-
-<style lang="sass">
-.model
-
-	&__header
-		border-bottom: 1px solid #eee
-
-	&__body
-		border-bottom: 1px solid #eee
-</style>
